% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{Corpus_Subset}
\alias{Corpus_Subset}
\title{Subsets of the MCFlow.}
\usage{
Corpus_Subset(artists = NA, emcees = NA, songs = NA, verses = 1:10,
  bpms = c(60, 150), years = 1980:2014, remove.hype = T,
  remove.rests = F)
}
\arguments{
\item{artists}{A vector of regex character strings, each representing an artist,
meaning the official commercial artist of the recording, to include in the subset.
 Defaults to all. Information about what artists are in the current corpus can be
 found in the \code{\link{CorpusData}} object.}

\item{emcees}{A vector of regex character strings, each representing an emcee to include
in the subset. Defaults to all. Information about what emcees are in the current corpus can be
  found in the \code{\link{CorpusData}} object.}

\item{songs}{A vector of regex character strings, each representing an song to include in
the subset. Defaults to all. The list of songs in the current corpus can be
  found in the \code{\link{CorpusData}} object.}

\item{verses}{A vector of integers, specifying which verses to include in the subset. Defaults to all, i.e. 1-10.}

\item{bpms}{A 2-length numeric vector indicating the range of accompaniment bpms to include in the susbet. Defaults to all, which is captured by the range (60, 150)}

\item{years}{A numeric vector of years to include in the sample. Defaults to all 1980-2014.}

\item{remove.hype}{A single logical value. If TRUE, all purely hype data records are removed from output.}

\item{remove.rests}{A single logical value. If TRUE, all rests are removed from output.}
}
\value{
Returns a subset of the \code{\link{FlowCorpus}} data object, which is a data.frame with
the same variables as any object of class \code{\link{Rap}}.
}
\description{
This function extracts a subset from the globally defined FlowCorpus data object.
}
\examples{
To get the data from Eminem rapped afte 2005:
Corpus_Subset(emcees='Eminem',years=2005:2014)
}

