% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{data}
\name{CorpusData}
\alias{CorpusData}
\title{Data about each song in MCFlow.}
\format{A data frame with 124 rows and 8 variables:
\describe{
  \item{Artist}{The name of the song's commercially releasing artist.}
  \item{Song}{The name of the song.}
  \item{ReleaseData}{The date the song was released.}
  \item{BillboardPeak}{The song's peak position on Billboard's \emph{Hot100}.}
  \item{BillboardPeakData}{The date the song reached it's Billboard peak.}
  \item{Emcees}{A character string representing the emcees who rap in each verse of the song, with each verse separated by a comma.}
  \item{N.Verses}{The number of verses in the song.}
  \item{BPM}{The beats per minute tempo of each verse in the song, in a character string separated by commas.}
}}
\usage{
CorpusData
}
\description{
General information about each song in the current Musical Corpus of Flow.
}
\keyword{datasets}

