% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{data}
\name{FlowCorpus}
\alias{FlowCorpus}
\title{MCFlow data in single object.}
\format{A data frame with 72,801 rows and 22 variables. The 22 variables are the same
as those found in an object of class \code{\link{Rap}}.}
\usage{
FlowCorpus
}
\description{
A single data frame containing all MCFlow data stacked on top of each other.
}
\keyword{datasets}

