% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{data}
\name{FlowData}
\alias{FlowData}
\title{MCFlow data, by song.}
\format{A list of objects of class Rap. Each object contains the complete transcription of
a song in MCFlow. The names(FlowData) are the artists' name and song name, seperated by a _
(underscore), with no spaces, commas, or apostrophes, in CamelCase. For instance: JaRule_LivinItUp.}
\usage{
FlowData
}
\description{
A list of objects of class \code{\link{Rap}}. Each object contains the complete transcription of
a song in MCFlow.
}
\keyword{datasets}

