% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{package}
\name{MCFlow}
\alias{MCFlow}
\alias{MCFlow-package}
\title{MCFlow: the Musical Corpus of Flow}
\description{
The MCFlow package includes the entire Musical Corpus of Flow,
a corpus of digital transcriptions of rap. The current version of
MCFlow contains 124 songs, by 47 artists.
}
\section{MCFlow datasets}{

Upon loading this package, three data objects are lazily loaded:
\enumerate{
\item The object \code{\link{FlowData}} is a list object containing each
of MCFlow's current 124 songs, each encoded as an object of class Rap.
The names(FlowData) correspond to each Artist_Song in the corpus
in CamelCase with no spaces, and no apostrophes.
Thus, Ja Rule's "Livin' It Up'" can be accessed with the command
FlowData[['JaRule_LivinItUp']].
\item The object \code{\link{FlowCorpus}} is a single data frame with
all the verses of all songs simply stacked on top of each other.
This is ideal for analyzing or generalizing across the whole corpus.
\item The object \code{\link{CorpusData}} is a data.frame object with
information about each of the 124 songs in the current MCFlow.
}
}

\section{MCFlow functions}{

A useful function in the MCFlow package is the function \code{\link{Corpus_Subset}}.
Corpus_Subset can be used to quickly find subsets of the \code{\link{FlowCorpus}} object
based on a variety of criteria. In addition, the MCFlow package includes
five visualization functions: \code{\link{plot.Rap}} (which makes flow diagrams of single
verses), \code{\link{plot_Speed}} (which plots histrograms of rapped syllable speeds),
\code{\link{plot_Metric}} (which plots the metric distribution of various syllable layers),
\code{\link{plot_Phrase}} (which plots the metric position of phrases in the meter), and
\code{\link{plot_Rhyme}} (which plots the 1st-order relationships between the metric positions
of antecedent/consequent rhymes).
}

