% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{plot.Rap}
\alias{plot.Rap}
\title{Flow diagramer}
\usage{
\method{plot}{Rap}(Rap, which.verse = 1, line.range = NULL,
  rhythmic.layers = "all", color.rhymes = T, metric.alignment = 32,
  plot.by.phrase = F, phrase.break = c("2", "3", "4"), IPA = F,
  cex = 0.8)
}
\arguments{
\item{Rap}{Object of class \code{\link{Rap}} (a list of data.frames).
The desired rap song; The data object \code{\link{FlowData}} contains \code{\link{Rap}} objects.}

\item{which.verse}{Numeric, length 1. Which verse of the rap to plot.}

\item{line.range}{Numeric vector of length two, specifying the first and last lines,
whether measures or phrases, to plot. If left NULL, then entire verse is plotted.}

\item{rhythmic.layers}{Character. Determines which rhythmic layers should be plotted.
Options are either 'all' 'stress' or 'rhymes'}

\item{color.rhymes}{Logical. Should rhymes be colored?}

\item{metric.alignment}{Metric distance at which successive lines are aligned in 32nd-note
units. 32 = one measure, 16 two beats, etc..}

\item{plot.by.phrase}{Logical. Plot by phrases instead of by measure?}

\item{phrase.break}{Vector of character strings. If plot.by.phrase==T, which break
annotations should be used to determine phrase boundaries? Options are '2', '3', or '4'.
First letters of phrases are also capitalized, whether or not plot.by.phrase==T.}

\item{IPA}{Print in the International Phonetic Alphabet instead of English orthography?}

\item{cex}{Size of text.}
}
\value{
none
}
\description{
This function makes a flow diagram of a given rap verse.
}
\examples{
To plot the second verse of 2pac's "California Love"
plot(FlowData[['2pac_CaliforniaLove]],which.verse=2,plot.by.phrase=T)
}
\keyword{Diagram}
\keyword{Flow}
\keyword{Rhyme}

