% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{plot_Metric}
\alias{plot_Metric}
\title{Plot syllables' metric distributions.}
\usage{
plot_Metric(Data = FlowCorpus, rhythmic.layers = c("unstressed", "stressed",
  "rhymed"), metric = 16, measures = 1, freq = F)
}
\arguments{
\item{Data}{The \code{\link{FlowCorpus}} data.frame or a subset of it.}

\item{rhythmic.layers}{Character vector, of length 1-3. Options are unstressed
stressed or rhymed...any combination of these three is acceptable}

\item{metric}{Numeric, the durational value to which all durations are rounded.
Defaults to 16th note. Other possibilites are 32 and 8 and 4.}

\item{measures}{The number of measures to show distributions across. Acceptable
options are .5, 1, and 2}

\item{freq}{Logical. If freq=T, total counts are shown, else proportion of total are shown.}
}
\value{
none
}
\description{
This plots the relative proportion of syllables to land in various metric positions.
}
\examples{
#To see the metric disitribution of rhymed syllables Young M.C.'s first verses.
plot_Metric(Corpus_Subset(artists='Young',verses=1),rhythmic.layers='rhymed')

}

