% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{plot_Phrase}
\alias{plot_Phrase}
\title{Plots metric position of phrases}
\usage{
plot_Phrase(Data = FlowCorpus, Wrap.beat = 1)
}
\arguments{
\item{Data}{The \code{\link{FlowCorpus}} data.frame or a subset of it.}

\item{Wrap.beat}{single numeric, between 1:16. Controls where the metric position of the
phrases at the bottom of the plot.}
}
\value{
none
}
\description{
This plots the metric position of phrases as a scatter of arrows across a metric grid.
}
\examples{
To see the metric distribution of all phrased in MCFlow rapped between 1992 and
1997:
plot_Phrase(Corpus_Subset(years=1992:1997))

}

