% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{plot_Rhyme}
\alias{plot_Rhyme}
\title{Plot first-order metric position of rhymes.}
\usage{
plot_Rhyme(Data = FlowCorpus, conditional = F, within.chain = T,
  rhyme.distance = 1, freq = F)
}
\arguments{
\item{Data}{The \code{\link{FlowCorpus}} data.frame or a subset of it.}

\item{conditional}{Single logical value. If true, conditional probabilities (given
antecedent rhyme) are shown. If false, the joint probability is shown.}

\item{within.chain}{Logical, length 1. If true, only rhyme relationships within a rhyme
chain (sounds that actually rhyme with each other) are counted.
Else, relationships between consecutive first-stressed-syllables in any rhyme are counted.}

\item{rhyme.distance}{Vector of numbers. Controls how close together in chain rhymes
must be. For instance, rhyme.distance==1 means that only adjacent rhymes in chain are
shown, while 2 means that everyother rhyme in chain is compared. Can be any or all values
 in the range 1:10.}

\item{freq}{Logical. If freq=T, total counts are shown else proportion of total are
shown (this only affects plot key). Conditional plots can't be freq (currently).}
}
\value{
none
}
\description{
This plots the relative probability (either joint or conditional) of consequent rhymes appearing on a given beat given an antecedent rhymes on a given beat.
}
\examples{
To plot Kurtis Blows 1st-rhyme patterns: plot_Rhyme(Corpus_Subset(artists="Kurtis"))

}

