% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{plot_Speed}
\alias{plot_Speed}
\title{Distribution of syllable by speed}
\usage{
plot_Speed(Data = FlowCorpus, unit = "SyllablesPerSecond",
  exclude.tails = c(0, ifelse(unit == "SyllablesPerSecond", 1, 0.97)),
  interquantile = 95, freq = F)
}
\arguments{
\item{Data}{The \code{\link{FlowCorpus}} data.frame or a subset of it.}

\item{unit}{Single character string, What kind of durations measures to plot. Options are SyllablesPerSecond, msIOIs, 16thIOIs}

\item{exclude.tails}{A 2-length numeric vector. This is used to screen out outliers. The first value in the vector is the lower (left) quantile, below which to exlude. The second value is the upper (right), above which to exclude.}

\item{interquantile}{A single number between 0 and 100. This marks the middle quantiles% of data under the plot.}

\item{freq}{Logical. If freq=T, total counts are shown, else proportion of total are shown.}
}
\value{
none
}
\description{
This plots a histogram of given Corpus subset's syllable speed.
}
\examples{
#To see the histogram of syllables per second delivered by Big Boi.
plot_Speed(Corpus_Subset(emcees='Big Boi'))

}

