% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{class}
\name{Rap}
\alias{Rap}
\title{The Rap S3 class}
\description{
The class of object of each MCFlow transcription.
A list of data.frames. Each data.frame represents one verse in the rap song.
The \code{\link{FlowData}} object contains all the current MCFlow Rap objects.
}
\section{Format}{

Each verse's data.frame includes a row for each syllable (and rest) in the verse, with
22 variables. The first nine variables are direct representations of the nine spines in each
MCFlow Humdrum-syntax text file. The next seven columns represent data that is stored in
Humdrum inerpretaton and reference records. The next six columns are various useful transformations
of the data in the first column.
\describe{
  \item{recip}{The MCFlow Humdrum-syntax **recipx spine. Tokens representing the duration of each event.}
  \item{stress}{The MCFlow Humdrum-syntax **stress spine. Tokens representing syllables' stress, as either 1 stressed, or 0 unstressed.}
  \item{tone}{The MCFlow Humdrum-syntax **tone spine. Tokens prosodic accents and boundary tones.}
  \item{breaks}{The MCFlow Humdrum-syntax **break spine. Tokens represent prosodic boundaries, from weakest 1 to strongest 4.}
  \item{melody}{The MCFlow Humdrum-syntax **melody spine. Tokens represnting prosodic (especially pitch) paralellism in flow.}
  \item{rhyme}{The MCFlow Humdrum-syntax **rhyme spine. Tokens "two-dimmesional" rhyme annotations, with roman letters marking syllable parallelism and parenthesis grouping rhymed syllables together.}
  \item{ipa}{The MCFlow Humdrum-syntax **ipa spine. Tokens represent rests (R) or International Phonetic Alphabet spellings of syllables' pronounciation.}
  \item{lyrics}{The MCFlow Humdrum-syntax **lyrics spine. Tokens rap's lyrics, with multisyllable words connected by - (dash) and acronyms and noun phrases connected by _ (underscore).}
  . (period) , (comma) ; (semicolon) and | (pipe) represent various syntactic boundaries. : (colon) and {} indicate semantic relationshins.
  \item{hype}{The MCFlow Humdrum-syntax **hype spine. Tokens hype (backing rap) in the song. recipx ipa and stress data jammed into each token.}
  \item{Year}{From MCFlow Humdrum-syntax reference records: the year (including decimal date) the song peaked on Billboard's \emph{Hot100} (or the year it was released if it never appeared on Billboard).}
  \item{Song}{From MCFlow Humdrum-syntax reference records: the song's name.}
  \item{Artist}{From MCFlow Humdrum-syntax reference records: the commcercially releasing artist.}
  \item{Emcee}{From MCFlow Humdrum-syntax comment data: The emcee rapping at any given time.}
  \item{BPM}{From MCFlow Humdrum-syntax interpretation data: The tempo in Beats Per Minute at each momment.}
  \item{Measure}{From MCFlow Humdrum-syntax barline data: The measure number (within verses) of each event. Measure 0 is a pickup measure.}
  \item{Section}{From MCFlow Humdrum-syntax interpretation data: The current section within the song (Verse1, Verse2 etc).}
  \item{dur.16}{Numeric. The event duration in 16th-notes.}
  \item{dur.ms}{Numeric. The event duration in milliseconds.}
  \item{ioi.16}{Numeric. The syllable's inter-onset-interval in 16th-notes.}
  \item{ioi.ms}{Numeric. The syllable's inter-onset-interval in milliseconds.}
  \item{syl.per.sec}{Numeric. The number of syllables rapped in the one-second window beginning with the current syllable.}
  \item{beat}{Numeric. The metric position (in 16th-notes) within the 4/4 measure.}
}
}

